requirements.txt
requests
beautifulsoup4
streamlit
streamlit-autorefresh
psycopg2-binary
google-genai
python-dotenv
plotly
langchain
langchain-chroma
chromadb
langchain-community
sentence-transformers
langchain-huggingface
torch